+++ frontend/src/App.tsx
@@ -4,7 +4,8 @@ import DashboardLayout from "@/routes/DashboardLayout";
 import LoginRoute from "@/routes/Login";
 import ProjectRoute from "@/routes/Project";
 import RegisterRoute from "@/routes/Register";
-import TimeTrackingRoute from "@/routes/TimetrackingRoute";
+import TimeRecordsRoute from "@/routes/TimeRecordsRoute";
+import TimeTrackingRoute from "@/routes/TimeTrackingRoute";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router";
 
 function App() {
@@ -22,6 +23,7 @@ function App() {
           <Route index element={<DashboardRoute />} />
           <Route path=":id" element={<ProjectRoute />} />
           <Route path=":id/timetracking" element={<TimeTrackingRoute />} />
+          <Route path=":id/timerecords" element={<TimeRecordsRoute />} />
         </Route>
         <Route path="/login" element={<LoginRoute />} />
         <Route path="/register" element={<RegisterRoute />} />