+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added reusable typography components
 - Added timeEntry routes to backend allowing to start, stop and retrieve working times
 - Added working integration of the time entries on a daily basis to the frontend
+- Added basic table displaying the time entries made so far
 
 ### Changed